delete book

+++ app/src/main/java/ru/ar4uk/bookstoreapp/ui/mainScreen/MainScreenViewModel.kt
@@ -3,7 +3,11 @@ package ru.ar4uk.bookstoreapp.ui.mainScreen
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.launch
 import ru.ar4uk.bookstoreapp.data.Book
 import ru.ar4uk.bookstoreapp.ui.mainScreen.bottom_menu.BottomMenuItem
 import ru.ar4uk.bookstoreapp.utils.firebase.FireStoreManager
@@ -16,19 +20,46 @@ class MainScreenViewModel @Inject constructor(
     val booksListState = mutableStateOf(emptyList<Book>())
     val isFavListEmptyState = mutableStateOf(false)
     val selectedBottomItemState = mutableStateOf(BottomMenuItem.Home.title)
+    var bookToDelete: Book? = null
 
-    fun getAllBooks() {
-        fireStoreManager.getAllBooks { books ->
-            booksListState.value = books
-            isFavListEmptyState.value = books.isEmpty()
+    private val _uiState = MutableSharedFlow<MainUiState>()
+    val uiState = _uiState.asSharedFlow()
+
+    private fun sendUiState(state: MainUiState) {
+        viewModelScope.launch {
+            _uiState.emit(state)
         }
     }
 
+    fun getAllBooks() {
+        sendUiState(MainUiState.Loading)
+
+        fireStoreManager.getAllBooks(
+            onBooks = { books ->
+                booksListState.value = books
+                isFavListEmptyState.value = books.isEmpty()
+                sendUiState(MainUiState.Success)
+            },
+            onFailure = {
+                sendUiState(MainUiState.Error(it))
+            }
+        )
+    }
+
     fun getAllFavsBooks() {
-        fireStoreManager.getAllFavsBooks { books ->
-            booksListState.value = books
-            isFavListEmptyState.value = books.isEmpty()
-        }
+        sendUiState(MainUiState.Loading)
+
+        fireStoreManager.getAllFavsBooks(
+            onBooks = { books ->
+                booksListState.value = books
+                isFavListEmptyState.value = books.isEmpty()
+
+                sendUiState(MainUiState.Success)
+            },
+            onFailure = {
+                sendUiState(MainUiState.Error(it))
+            }
+        )
     }
 
     fun getBooksFromCategory(category: String) {
@@ -37,10 +68,19 @@ class MainScreenViewModel @Inject constructor(
             return
         }
 
-        fireStoreManager.getAllBooksFromCategory(category) { books->
-            booksListState.value = books
-            isFavListEmptyState.value = books.isEmpty()
-        }
+        sendUiState(MainUiState.Loading)
+
+        fireStoreManager.getAllBooksFromCategory(
+            category,
+            onBooks = { books->
+                booksListState.value = books
+                isFavListEmptyState.value = books.isEmpty()
+                sendUiState(MainUiState.Success)
+            },
+            onFailure = {
+                sendUiState(MainUiState.Error(it))
+            }
+        )
     }
 
     fun onFavClick(book: Book, isFavState: String) {
@@ -53,16 +93,25 @@ class MainScreenViewModel @Inject constructor(
         isFavListEmptyState.value = booksListState.value.isEmpty()
     }
 
-    fun deleteBook(
-        book: Book
-    ) {
+    fun deleteBook() {
+        if (bookToDelete == null) return
+
         fireStoreManager.deleteBook(
-            book,
+            bookToDelete!!,
             onDeleted = {
                 booksListState.value = booksListState.value.filter {
-                    it.id != book.id
+                    it.id != bookToDelete!!.id
                 }
+            },
+            onFailure = {
+                sendUiState(MainUiState.Error(it))
             }
         )
+    }
+
+    sealed class MainUiState {
+        data object Loading: MainUiState()
+        data object Success: MainUiState()
+        data class Error(val message: String): MainUiState()
     }
 }
